+++ blog/admin/posts/list.php
@@ -43,7 +43,7 @@
             <td><?= $row['CATEGORY_TITLE'] ?></td>
             <td><?= $row['IS_PUBLISHED'] == 1 ? 'Yes' : 'No' ?></td>
             <td><a href="edit.php?id=<?= $row['ID'] ?>">EDIT</a></td>
-            <td><a href="delete.php?id=<?= $row['ID'] ?>">DELETE</a></td>
+            <td><a href="delete.php?id=<?= $row['ID'] ?>" onclick="return confirm('Are you sure that you want to delete post titled: <?= $row{'TITLE'} ?>')">DELETE</a></td>
         </tr>
         <?php endwhile; ?>
     </table>
